[ticket/12099] Correctly fix the path when performing AJAX requests
If the current request is a AJAX we need to fix the paths.
We need to get the root path based on the Referer, so we can use
the generated URLs in the template of the Referer. If we do not
generate the relative path based on the Referer, but based on the
currently requested URL, the generated URLs will not point to the
intended locations:
	Referer				desired URL			desired relative root path
	memberlist.php		faq.php				./
	memberlist.php		app.php/foo/bar		./
	app.php/foo			memberlist.php		../
	app.php/foo			app.php/fox			../
	app.php/foo/bar		memberlist.php		../../
	../page.php			memberlist.php		./phpBB/
	../sub/page.php		memberlist.php		./../phpBB/

PHPBB3-12099